Initial Fork
Cleaned up the files more to my liking.

in Program.cs I added string CurrentDirectory = AppDomain.CurrentDomain.BaseDirectory;
so that we can get our directory a little easier.

Added check.py for elevation testing.

The binary seems to be getting defended. So this method probably will no longer work. This is likely just an archive and will not be maintained.